Hopefully produce desired database
This code is running as we speak but with the 15minute sleep cycles, it
will take a few hours to complete. Unfortunately it wont print the
messages to console during the program, it will only do this once the
code has finished compiling, so as of now Im snot sure if it will
produce the desired database of 2 columns. One for the rugby player and
the other for that rugby players friends