Failed attempt to create database
'Connection aborted.', error(10054, 'An existing connection was forcibly
closed by the remote host')) is the error I got but the main problem is
that the rate limit was exceeded too often. If you look at the console
messages from this compilation you can see how many times the sleep
function was used. Also the console messages wont show while the program
is running which is wasting alot of time.